DROID-3965 mapping
